Move plugin funcs to plugins
